Add missing apk build steps
Looks like some build steps, like actually building the "launcher"
source code, were left of the script and have now been added. :)

Also it turns out that for the build to finish successfully,
paths to the Android NDK & SDK and to Qt Android qmake need to be
provided, so and the script now expects the script "env.sh" to exist in the
scripts folder, which exports the needed paths as environmental
variables.

So like this, it is now possible to create a new modRana apk directly
from command line in a rather automatic fashion. :)